wip: create frame.py (include class Frame) and viewport.py (include methods find_area and set_view_distance)
